% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive_scrub.R
\name{passive_scrub}
\alias{passive_scrub}
\title{Audio file scrubbing for (potentially) multiple night passive recording.}
\usage{
passive_scrub(calls = NULL)
}
\arguments{
\item{calls}{optional character string to specify the path to the BCID classification
output file; default (`NULL`) allows the user to navigate to the file via a dialog box}
}
\description{
This function uses the classification output from Bat Call Identification (BCID;
 \url{http::/www.batcallid.com}) software to scrub (move) suspected noise files into
 a new subdirectory.  Unlike mobile data, which is typically collected in a single night,
 passive data are often collected over multiple nights at a given location.  This function
 is designed to handle multiple nights of recordings in the BCID output, unlike
}

