% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_MABM_route.R
\name{plot_MABM_route}
\alias{plot_MABM_route}
\title{Plot MABM route and, if requested, bat calls detected along the route.}
\usage{
plot_MABM_route(type = "roadmap", spp = TRUE, bats = c("all", "good",
  "bad", "none"), labels = TRUE, bad_gps = 5)
}
\arguments{
\item{type}{character string providing map theme. Reasonable choices include "terrain",
"terrain-background", "satellite", "roadmap" (the default), and "hybrid" (google maps).}

\item{spp}{logical indicating whether detections should be categorized by species;
defaults to TRUE.}

\item{bats}{character string indicating which, if any, bat calls to display along the
MABM route.  Options available include "all" (default), "good", "bad", and "none".
The distinction between "good" and "bad" calls depends on the user-specified
\code{bad_gps} argument.}

\item{labels}{logical (default = TRUE) indicating whether every third detection
should be labeled with a time stamp to assist in visualizing the route. Ignored if the
user opts not to add detections to the route.}

\item{bad_gps}{integer indicating the absolute maximum number of seconds between a
detected bat call and the nearest (in time) GPS location fix the user will consider as
a "good" (usually interpreted "usable") georeference for a given bat call.  All calls
greater than this value will be recorded as having a "bad" georeference. The actual
time difference between a bat detection and the closest GPS location is noted in the
output under the column heading 'GPS_diff'; positive (negative) values indicate the
detection occurred after (before) the associated GPS location.}
}
\value{
a ggmap object (a classed raster object with a bounding box attribute)
}
\description{
This function plots a MABM route (from a shapefile created with \code{\link{MABM_route}})
and, if requested (with the \code{bats} argument), bat calls (by species identification)
detected along the route. To work properly this function requires an internet connection.
}
\details{
With the \code{bats} argument, the user can specify whether all detected calls should be
displayed or only those that meet (or do not meet) a user-specified criterion for a "good"
georeferenced call. This criterion is specified as the number of seconds between a detected
call and the closest (in time) GPS fix.

Species classifications can be turned off using \code{spp = FALSE}.

By default, every third bat detection (when plotted) is labeled with its associated time
stamp and the route is shaded by time (minutes) elapsed from the first GPS fix to help
visualize the route.  These can be suppressed with \code{labels = FALSE}.
}

