% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_MABM_root.R
\name{set_MABM_root}
\alias{set_MABM_root}
\title{Interactively set the root directory of all MABM data}
\usage{
set_MABM_root(dir = NULL)
}
\arguments{
\item{dir}{character string indicating the base directory containing MABM station
related data (i.e., MABM route directories with call files, shapefiles, and annual reports).}
}
\description{
This function allows you to specify the root directory containing all MABM routes and
 associated data, significantly easing the navigation burden when processing BCID output
 using `MABM_route`.  Set this once at the beginning of a session (see example).
}
\examples{
\dontrun{
#' # Browse to specify MABM root directory
set_MABM_root()

# Speficy MABM root directory directly
set_MABM_root("C:/MABM")
}
}
