% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_MABM_route.R
\name{plot_MABM_route}
\alias{plot_MABM_route}
\title{Plot MABM route and, if requested, bat calls detected along the route.}
\usage{
plot_MABM_route(bad_gps = 5, basemap = c("terrain", "aerial"), gps = NULL)
}
\arguments{
\item{bad_gps}{integer indicating the absolute maximum number of seconds between a
detected bat call and the nearest (in time) GPS location fix the user will consider as
a "good" (usually interpreted "usable") georeference for a given bat call.  All calls
greater than this value will be recorded as having a "bad" georeference. The actual
time difference between a bat detection and the closest GPS location is noted in the
output under the column heading 'GPS_diff'; positive (negative) values indicate the
detection occurred after (before) the associated GPS location.}

\item{basemap}{character string that allows the user to select between a
terrain ("terrain"; default) or aerial imagery ("aerial")}

\item{gps}{character string that gives user the option to specify the path (full or
relative) to the GPS point shapefile rather than using a dialog box}
}
\value{
an HTML widget object
}
\description{
This function plots a MABM route (from a shapefile created with \code{\link{MABM_route}})
and, if requested (with the \code{bats} argument), bat calls (by species identification)
detected along the route. To work properly this function requires an internet connection.
}
\details{
With the \code{bats} argument, the user can specify whether all detected calls should be
displayed or only those that meet (or do not meet) a user-specified criterion for a "good"
georeferenced call. This criterion is specified as the number of seconds between a detected
call and the closest (in time) GPS fix.

Species classifications can be turned off using \code{spp = FALSE}.

By default, every third bat detection (when plotted) is labeled with its associated time
stamp and the route is shaded by time (minutes) elapsed from the first GPS fix to help
visualize the route.  These can be suppressed with \code{labels = FALSE}.
}

