% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scrub_noise.R
\name{scrub_noise}
\alias{scrub_noise}
\title{Audio file scrubbing expected noise files one or more nights of recording.}
\usage{
scrub_noise(mult_folder = TRUE, BCID = NULL)
}
\arguments{
\item{mult_folder}{logical (default = `TRUE`) indicating whether Anabat files can be
found in separate nightly folders (use `TRUE`) or in the top-level directory (i.e.,
the same directory as the BCID .xls file; use 'FALSE')}

\item{BCID}{optional character string to specify the path to the BCID classification
output file; default (`NULL`) allows the user to navigate to the file via a dialog box}
}
\description{
This function uses the classification output from Bat Call Identification (BCID;
 \url{http::/www.batcallid.com}) software to scrub (move) suspected noise files into
 a new subdirectory.  This function
 is designed to handle multiple nights of recordings in the BCID output.
}

